Shared functions enhancement
