+++ packages/core/src/rooms/RoomSessionPlayback.test.ts
@@ -13,16 +13,19 @@ describe('RoomSessionPlayback', () => {
       instance = createRoomSessionPlaybackObject({
         store: configureJestStore(),
         emitter: new EventEmitter<RoomSessionPlaybackEventsHandlerMapping>(),
+        payload: {
+          //@ts-expect-error
+          playback: {
+            id: 'c22d7223-5a01-49fe-8da0-46bec8e75e32',
+          },
+          room_session_id: 'room-session-id',
+        },
       })
       // @ts-expect-error
       instance.execute = jest.fn()
     })
 
     it('should control an active playback', async () => {
-      // Mock properties
-      instance.id = 'c22d7223-5a01-49fe-8da0-46bec8e75e32'
-      instance.roomSessionId = 'room-session-id'
-
       const baseExecuteParams = {
         method: '',
         params: {
