+++ packages/core/src/rooms/index.ts
@@ -13,6 +13,5 @@ export * from './methodsRT'
 export * from './RoomSessionRecording'
 export * from './RoomSessionRTRecording'
 export * from './RoomSessionPlayback'
-export * from './RoomSessionRTPlayback'
 export * from './RoomSessionStream'
 export * from './RoomSessionRTStream'
